CI: Use class factories for common DOS definitions
